remove more field files; add ol-test-utilities
