+++ frontends/api/package.json
@@ -17,10 +17,10 @@
     "@faker-js/faker": "^7.3.0",
     "@testing-library/react": "^14.1.2",
     "jest-when": "^3.6.0",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "ol-test-utilities": "workspace:*"
   },
   "dependencies": {
-    "@lukemorales/query-key-factory": "^1.3.2",
-    "ol-utilities": "workspace:*"
+    "@lukemorales/query-key-factory": "^1.3.2"
   }
 }
