+++ frontends/mit-open/src/components/RestrictedRoute/RestrictedRoute.test.tsx
@@ -6,7 +6,7 @@ import {
 } from "../../test-utils"
 import RestrictedRoute from "./RestrictedRoute"
 import { ForbiddenError, Permissions } from "@/common/permissions"
-import { allowConsoleErrors } from "ol-utilities/test-utils"
+import { allowConsoleErrors } from "ol-test-utilities"
 
 test("Renders children if permission check satisfied", () => {
   renderWithProviders(
