+++ frontends/mit-open/src/page-components/LearningResourceCardTemplate/LearningResourceCardTemplate.test.tsx
@@ -2,8 +2,8 @@ import React from "react"
 import { render, screen } from "@testing-library/react"
 import { assertInstanceOf, resourceThumbnailSrc } from "ol-utilities"
 import LearningResourceCardTemplate from "./LearningResourceCardTemplate"
-import { makeImgConfig } from "ol-utilities/factories"
-import { allowConsoleErrors } from "ol-utilities/test-utils"
+import { makeImgConfig } from "ol-utilities/test-utils/factories"
+import { allowConsoleErrors } from "ol-test-utilities"
 import user from "@testing-library/user-event"
 import * as factories from "api/test-utils/factories"
 import { ResourceTypeEnum } from "api"
