+++ frontends/mit-open/src/pages/ErrorPage/ErrorPage.test.tsx
@@ -5,7 +5,7 @@ import { renderRoutesWithProviders, renderTestApp } from "../../test-utils"
 import type { TestAppOptions } from "../../test-utils"
 import ErrorPage from "./ErrorPage"
 import { setMockResponse, mockAxiosInstance as axios } from "api/test-utils"
-import { allowConsoleErrors } from "ol-utilities/test-utils"
+import { allowConsoleErrors } from "ol-test-utilities"
 import RestrictedRoute from "@/components/RestrictedRoute/RestrictedRoute"
 import { Permissions } from "@/common/permissions"
