+++ frontends/mit-open/src/test-utils/setupJest.ts
@@ -29,37 +29,6 @@ afterEach(() => {
   window.SETTINGS = _createSettings()
 })
 
-/**
- * We frequently spy on these, so let's just do it once.
- */
-jest.mock(
-  "@/page-components/LearningResourceCardTemplateLegacy/LearningResourceCardTemplateLegacy",
-  () => {
-    const actual = jest.requireActual(
-      "@/page-components/LearningResourceCardTemplateLegacy/LearningResourceCardTemplateLegacy",
-    )
-    return {
-      __esModule: true,
-      ...actual,
-      default: jest.fn(actual.default),
-    }
-  },
-)
-
-jest.mock(
-  "@/page-components/ExpandedLearningResourceDisplay/ExpandedLearningResourceDisplay",
-  () => {
-    const actual = jest.requireActual(
-      "@/page-components/ExpandedLearningResourceDisplay/ExpandedLearningResourceDisplay",
-    )
-    return {
-      __esModule: true,
-      ...actual,
-      default: jest.fn(actual.default),
-    }
-  },
-)
-
 jest.mock("@/page-components/LearningResourceCard/LearningResourceCard", () => {
   const actual = jest.requireActual(
     "@/page-components/LearningResourceCard/LearningResourceCard",
