+++ frontends/ol-ckeditor/package.json
@@ -32,7 +32,9 @@
     "@ckeditor/ckeditor5-theme-lark": "^40.1.0",
     "@ckeditor/ckeditor5-ui": "^40.1.0",
     "axios": "^0.27.2",
-    "classnames": "^2.3.2"
+    "classnames": "^2.3.2",
+    "ol-components": "workspace:*",
+    "ol-utilities": "workspace:*"
   },
   "devDependencies": {
     "@testing-library/react": "^14.1.2"
