--- frontends/ol-common/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "ol-common",
-  "packageManager": "yarn@3.7.0",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "dependencies": {
-    "@mitodl/course-search-utils": "^2.3.2"
-  }
-}
