--- frontends/ol-common/src/interfaces/search-ui.ts
@@ -1,5 +1,0 @@
-export type EmbedlyConfig = {
-  key: string
-  width: number
-  height: number
-}
