+++ frontends/ol-components/package.json
@@ -11,12 +11,12 @@
     "@dnd-kit/utilities": "^3.2.1",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mitodl/course-search-utils": "^2.3.2",
     "@mui/icons-material": "^5.14.19",
     "@mui/lab": "^5.0.0-alpha.150",
     "@mui/material": "^5.14.15",
     "lodash": "^4.17.21",
     "nuka-carousel": "^7.0.0",
+    "ol-test-utilities": "workspace:*",
     "ol-utilities": "workspace:*",
     "rc-tooltip": "^6.1.2",
     "react-router": "^6.19.0",
