+++ frontends/ol-test-utilities/jest.config.ts
@@ -1,0 +1,8 @@
+import type { Config } from "@jest/types"
+import baseConfig from "../../jest.jsdom.config"
+
+const config: Config.InitialOptions = {
+  ...baseConfig,
+}
+
+export default config
