+++ frontends/ol-test-utilities/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "ol-test-utilities",
+  "packageManager": "yarn@3.7.0",
+  "exports": {
+    ".": "./src/index.ts",
+    "./filemocks/*": "./src/filemocks/*"
+  },
+  "dependencies": {
+    "@faker-js/faker": "^7.3.0",
+    "@testing-library/react": "^14.1.2",
+    "css-mediaquery": "^0.1.2",
+    "tiny-invariant": "^1.3.1"
+  }
+}
