+++ frontends/ol-test-utilities/src/domQueries.ts
@@ -1,37 +1,6 @@
 import { buildQueries, within } from "@testing-library/react"
 import type { GetErrorFunction } from "@testing-library/react"
-import mediaQuery from "css-mediaquery"
-import { assertInstanceOf } from "../predicates"
-
-/**
- * Create a mock mediaQuery functiton. Matches are determined by comparison with
- * the provided values.
- *
- * For example:
- * ```
- * const mediaQuery = createMatchMediaForJsDom({ width: "100px" })
- * // Will return false; width is "100px", which is below minimum
- * mediaQuery("(min-width: 120px)").match
- * ```
- *
- * See
- *  - https://jestjs.io/docs/manual-mocks#mocking-methods-which-are-not-implemented-in-jsdom
- *  - https://mui.com/material-ui/react-use-media-query/#testing
- */
-const createMatchMediaForJsDom = (
-  values: Partial<mediaQuery.MediaValues>,
-): ((query: string) => MediaQueryList) => {
-  return (query: string) => ({
-    matches: mediaQuery.match(query, values),
-    media: query,
-    addListener: jest.fn(),
-    removeListener: jest.fn(),
-    addEventListener: jest.fn(),
-    removeEventListener: jest.fn(),
-    onchange: null,
-    dispatchEvent: jest.fn(),
-  })
-}
+import invariant from "tiny-invariant"
 
 /**
  * Get all <dd> elements whose corresponding <dt> element matches the given term.
@@ -92,21 +61,8 @@ const getDescriptionFor = (el: HTMLElement) => {
   }
   // eslint-disable-next-line testing-library/no-node-access
   const errEl = document.getElementById(errId)
-  assertInstanceOf(errEl, HTMLElement)
+  invariant(errEl instanceof HTMLElement)
   return errEl
-}
-
-/**
- * We use [jest-fail-on-console](https://www.npmjs.com/package/jest-fail-on-console)
- * to fail on console errors/warnings.
- *
- * This function allows us to temporarily disable that behavior for a test,
- * e.g., to test API error handling.
- */
-const allowConsoleErrors = () => {
-  const consoleError = jest.spyOn(console, "error").mockImplementation()
-  const consoleWarn = jest.spyOn(console, "warn").mockImplementation()
-  return { consoleError, consoleWarn }
 }
 
 export {
@@ -117,7 +73,4 @@ export {
   findAllByTerm,
   findByTerm,
   getDescriptionFor,
-  createMatchMediaForJsDom,
-  allowConsoleErrors,
 }
-export { default as ControlledPromise } from "./ControlledPromise"
