+++ frontends/ol-test-utilities/src/filemocks/filemock.js
@@ -1,0 +1,1 @@
+// Used by jest to mock generic file imports
