+++ frontends/ol-test-utilities/src/filemocks/svgmock.js
@@ -1,1 +1,2 @@
+// Used by jest to mock SVG imports
 module.exports = "<svg><!-- Mock SVG --></svg>"
