+++ frontends/ol-utilities/package.json
@@ -4,11 +4,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./assets/*": "./assets/*",
-    "./test-utils": "./src/test-utils/index.ts",
-    "./factories": "./src/factories.ts",
-    "./deprecated": "./src/deprecated/index.ts",
-    "./test-utils/filemock": "./src/test-utils/filemock.js",
-    "./test-utils/svgmock": "./src/test-utils/svgmock.js"
+    "./test-utils/factories": "./src/test-utils/factories.ts"
   },
   "peerDependencies": {
     "react": "^18.2.0",
@@ -19,6 +15,7 @@
     "@dnd-kit/sortable": "^7.0.2",
     "@dnd-kit/utilities": "^3.2.1",
     "@faker-js/faker": "^7.3.0",
+    "api": "workspace:*",
     "classnames": "^2.3.2",
     "decimal.js-light": "^2.5.1",
     "html-entities": "^2.4.0",
