--- frontends/ol-utilities/src/factories.ts
@@ -1,1 +1,0 @@
-export * from "./test-utils/factories/factories"
