+++ frontends/ol-utilities/src/index.ts
@@ -9,6 +9,5 @@ export * from "./learning-resources/learning-resources"
 export * from "./strings/html"
 export * from "./predicates"
 export * from "./hooks"
-export * from "./test-utils/factories/interfaces"
 export * from "./querystrings"
 export * from "./lib"
