+++ frontends/ol-widgets/package.json
@@ -9,6 +9,7 @@
     "formik": "^2.4.5",
     "lodash": "^4.17.21",
     "ol-ckeditor": "workspace:*",
+    "ol-components": "workspace:*",
     "ol-utilities": "workspace:*",
     "react-markdown": "^6.0.3",
     "remark-gfm": "^1.0.0",
